requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
azure-ai-contentsafety==1.0.0b1
azure-ai-formrecognizer==3.3.0
azure-ai-language-conversations==1.1.0
azure-ai-textanalytics==5.3.0
azure-common==1.1.28
azure-core==1.29.3
blinker==1.6.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
cryptography==41.0.3
Flask==2.3.3
frozenlist==1.4.0
idna==3.4
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
msal==1.23.0
msrest==0.7.1
multidict==6.0.4
numpy==1.25.2
oauthlib==3.2.2
openai==0.28.0
opencv-python==4.8.0.76
pycparser==2.21
PyJWT==2.8.0
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
tqdm==4.66.1
typing_extensions==4.7.1
urllib3==2.0.4
Werkzeug==2.3.7
yarl==1.9.2
